로그아웃 기능 및 책 조회 기능 개선

+++ src/main/java/com/bookshelf/book/domain/Book.java
@@ -29,7 +29,7 @@ public class Book {
     @JoinColumn(name = "member_id")
     private Member member;
 
-    private Long likes;
+    private Integer likes;
 
     private boolean bookMark;
